refactor(app): split bootstrap from runtime service
